+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.0] - 2026-07-25
+
+### Added
+
+- Added proper support for subfolders.
+
+### Fixed
+
+- Fixed certain repair recipes not being hidden from EMI.
+- Fixed certain types of recipe replacements not properly reflecting in EMI.
+
+### Removed
+
+- Removed legacy config migrator.
+
 ## [3.0.1] - 2026-07-25
 
 ### Changed